+++ src/typing_ex/view/page.clj
@@ -9,7 +9,7 @@
    #_[taoensso.timbre :as timbre]
    [typing-ex.plot :refer [scatter]]))
 
-(def ^:private version "1.7.0")
+(def ^:private version "1.8.0")
 
 (defn page [& contents]
   [::response/ok